Avoid checking FFI constants for usability
In #4349 we saw that this logic does not load properly when native
function calls have been disabled, due to FFI partially loading
and defining an FFI module that was not fully functional. This
is why the previous code here checked for the existence of
FFI::Config rather than just FFI.

However this is still error-prone, so I am restructuring this
logic to do all FFI juggling within the same rescue block as the
require of 'ffi'. If that require raises LoadError or if the
required functions are not present, it immediately falls back on
defining the NotImplemented version of File.symlink.

The same change has been done to the Solaris File#flock logic
below.