Handle varargs frame on arity error in jitted code
Jitted code handles arity-checking differently than in the
interpreter. Where the interpreter always pushed a backtrace
element before the arity check, the JVM does the arity-check in a
separately-generated varargs wrapper (if the method is specific-
arity) or before setting a line number (if the method is variable-
arity). This has the effect of either omitting the error frame
altogether (specific) or having the line number be -1 (variable).

This commit makes the following modifications to allow the arity
error line to be shown in jitted code:

* Set the current line number immediately upon creating the method
  so that it is set in the JVM before the arity check.
* Process jitted method names with the varargs marker if they were
  not preceded by the actual method body. The varargs-marked
  method can only appear on its own or immediately after the real
  jitted method body.

This fixes all known cases where the argument error line is
omitted or not set properly in the JVM trace.

In addition, this patch moves the management of the current line
number into the IRBytecodeAdapter, rather than maintaining it in
the JVMVisitor instance.

Fixes #6335.

This may also improve the case described in #5103, since all JVM
frames will have proper line numbers at the point of error.