More agressively kill dead code
LocalOptimizationPass will propagate values.  For example:

```ruby
_tmp = true
unless _tmp
 puts "Never"
end
```

In this case by the end of this pass we will see a BB end with a
b_false(true, some_label).  This clearly cannot happen.  If we eliminate
this instr then we should also be able to lop off the bb(s) representing
whatever is within the unless conditional.  In this snippet the entire
snippet should be dead code except _tmp = true.

To fix this I added simplifyBranch to BranchInstr.  This gives us the
ability to swap out a branch with a noop or a jump (in case where something
is unconditionally true).

After we replace the instr with a new one we need to reexamine the basicblock
it is in and correct any edge placement.