Fix failing tests.
The basic problem was I changed this branch/PR to save the original
returnBBs as a field in the CFG.  When I remove a BB I have to also
update the returnBB list as well.

Second fix was to remove incoming edges from BBs which no longer are
reachable from the BBs by using the directed graphs removeEdge method.
This is quite a bit cleaner looking as well.