Fix regression introduced in recent dead code PR
I made the nil operand into a static field forgetting that it caches
a live value.  This meant during JSR223 testing which is setting up
multiple ruby runtimes were all using this static field and newer
runtimes would make their own nil but then be comparing with this
immortal cached nil.  This somehow ended up making tests fail because
it was getting an old 'main' object and a new 'main' object and rightfully
realizing they are not the same receiver object.

Nil is a field on IRManager again and simplifyBranch now passes in
fic so it can get to manager and get that mortal nil.