Expand super logic to handle Java proxy supers
Super method invocations from JRuby must use the generated super
shims in order to invoke the override at the correct level in the
class hierarchy. The previous logic fell back on standard
reflective invocation, which always starts at the bottom. This led
to the issues in #6718 where the super call chain would get stuck
in a recursive loop and eventually blow the call stack. This
change moves the invocation logic closer to the Ruby super logic,
allowing us to retrieve the appropriate super shim method when
performing the invocation.

Fixes #6718