Modify expected jar sizes for 9.3
jruby-core: The value of 10 was for the unshaded jar in JRuby
9.2 and when the shading of jruby-core was fixed it no longer
matched. The jar is actually *smaller* in JRuby 9.3 due to things
like lambdas removing inner classes.

jruby-complete: Several libraries that were loose in stdlib are
now gems, and some of these libraries have increased in size. The
9.2 jar did not appear to include the JNR x86_64 and aarch64 asm
libraries. Bundler is now included with JRuby and appears in the
complete jar.