Revert to ostruct 0.1.0 to avoid alias warnings
Newer ostruct includes code that aliases all core methods,
including several Kernel methods that access the caller's frame.

See ruby/ostruct#30 and jruby/jruby#6831

We are punting on this until the ostruct issue is addressed. This
change reverts to an earlier version of ostruct that does not have
the offending code.

A user manually updating to the newer ostruct will run into these
warnings again, so the issue remains unresolved. At least we will
not warn for 9.3.0.0.