Exclude test that does not correspond to gem
The released 0.1.0 gem, which is what 2.6.8 claims to ship, does
not have the changes in this test. This is due to CRuby having
made custom changes to their local copy of ostruct in the CRuby
repository, and not releasing that custom version.

See https://bugs.ruby-lang.org/issues/18169

We will exclude this test for now.