Avoid alias warning if target has same reqs
We warn when aliasing a method that requires the caller's frame
due to the fact that we use those names to force methods to push
a frame when they would not otherwise need it. However, if the
target name represents one that has the same or a superset of
frame requirements, we can safely perform the alias, knowing that
the new name is already flagged globally.

This patch checks whether the target name has the same or a
superset of frame requirements and skips the warning.

The defineAliases(..., List) form is deprecated as it does not
appear to be in use by any versioned or generated code.

The searchForAliasMethod method no longer performs the warning
check, since the deprecated method would need to perform it for
each entry and a new method checkAliasFrameAccesses does the check
and warn.

This relates to ruby/ostruct#30, but does not directly work around
any of the issues there. It would comprise part of a workaround if
we also pre-marked the instance_eval! and instance_exec! methods,
but that is not currently planned (and we would rather not).