Actually implement logic of associateEncoding in Ripper.
When strings are made from a parser encoding of US-ASCII and a new string contains
characters which are not explicitly UTF-8 then the encoding should be made BINARY.
The logic before did this but it called a method associateEncoding which was
unimplemented.

This logic is so obscure it took 6 years to get hit by anyone.  It is also so
obscure that so far I have been unabled to make a reproduction.