Fix handling of ENXIO error during resource open
This triggered a failure in the io/wait tests when it expected a
nonblocking write-only open call against an unread fifo to raise
ENXIO. We instead raised one of our internal anonymous errors,
preventing the logic in this test from retrying:

https://github.com/ruby/io-wait/blob/0c73ebcf5d75b0568405f5863eefaffba0ec7a0e/test/io/wait/test_io_wait_uncommon.rb#L23

The changes here promote all unraised errno exceptions to use
ErrnoException, which propagates the errno object as well as the
name and path. This allows upstream error handling to raise the
proper Ruby error.