Fix KeyError behavior with no arguments
After JRuby 9.3, attempting to raise a KeyError with no arguments would
fail with an ArgumentError. This commit fixes this behavior to match
MRI, and updates the `key` and `receiver` methods to raise if they are
unset.

Prior to this, the relevant MRI tests for this functionality were not
running in CI, and when ran manually 3 tests were failing. Add the
`test_key_error.rb` file to the `mri.code` suite. After this change, all
tests from `test_key_error.rb` pass.

Fixes #6860.