Fix #6859.  respond_to? regression with refinements.
respond_to? calls searchMethod and that code path ends up passing
the static scope as null.  RubyModule.searchWithCache notices we
are dealing with a refined method yet it has no static scope so it
just passes by the refinement logic.

The fix is to get current static scope based on the notion current
is the lexical location anything would be searching at the moment.
Getting current context as the fix is a little undesirable but pushing
context through all these paths would be a big change.