Beginnings of a new sprintf parser/interpreter.
The goal of this new design is to make it much simpler to modify
and update when we find a bug or have to add a new identifier.  A
secondary longer goal is to decouple processing of the format string
from the generation of the final string when executing sprintf.  The
value of this decoupling is we can push the format parsing phase into
the sprintf callsite so we only have to process the format once.  In
fact once we get to that point we can potentially just bytecode generate
the string building piece.

The current point of this work is we can completely parse the format
string into a sequence of tokens.  It is done sans bugs or decisions to
add/change any fields based on work on the interpreter.

At this point to use it you need to set the env var SPRINTF=anything
to enable new runtime and it will only work with %i, %d, and %u.

Future work will be to keep adding new directives but not be trying
very hard to make one method process to many directives.  In the case
of the three (i,d,u) 'i' is an alias of 'd' and 'u' processes unsigned
values a little differently when setting up the value.   Any more complexity
and u would have been its own method.