Add an option to turn off "foo".freeze optz
Officially, the .freeze optimization should still see when the
freeze method is replaced and actually call it, but the use cases
for this are so rare (we know of zero) we are instead just adding
an option to turn off the optimization.

This is a workaround for the missing behavior described in #2156.