Honor the requested address even given hostname
Previously if the hostname was given it was looked-up rather than
being used with the requested address. The new logic will use the
hostname if alone or hostname plus address otherwise.

Also included in this patch, the array accesses now use `entry`
which will return a nil for missing items. Previous use of `eltOk`
would raise a ConcurrencyError if the array was not long enough.