Allow immediates and frozen objects as keys
Note that some objects that are "immediate" in CRuby, such as
fixnums and flonums, cannot truly be idempotent on JRuby. This
means their use as keys will have unexpected behavior, due to the
potential for two equal values to actually be different object
identities.

Hopefully this does not become a common use case.