Start wiring up explicit keyword args
This introduces a new type, KeywordHash, and logic that recognizes
this (and only this) as the mechanism by which we pass keyword
arguments on the stack.

There's a lot of hackery here to make sure duped copies of KWHash
reman KWHash, arity ignores it, keyword receives expect it, and
so on. It's messy, but it is starting to pass more of the 3.0
kwarg specs than when I started.