Exclude broken parse test imported from json
This test passes to DateTime.parse a keyword argument
"create_additions" that is unknown to DateTime, but because CRuby
only checks for the "limit" option it expects, this does not cause
any error. In JRuby, where we still have this code in Ruby, the
unknown keyword is explicitly rejected. We exclude this for now
and will fix the bad test in json lib.