Fix Argf#codepoints++
Argf#codepoints without a block was not even giving back codepoints but
was giving out lines.  Once this was fixed the code in RubyIO which would
produce these codepoints was not passing in a valid encoding to
preciseLength which makes me think perhaps this is the only method which
uses this codepath.  Then the logic to advance files in the block version
of Argf#codepoints was not advancing through the files.  A pretty strange
cascade of bugs.

The second thing this commit does is reformulate next_p from using {-1, 0, 1}
to an enum which makes it easier to understand what is happening.  While doing
this I realized we had a dead code path checking argv.isNil().  argv is
unconditionally a RubyArray.