Array joining on something which clears array should work
The struct was being accessing an element and clearing the
array which was creating a negative size for the result string.
The logic added actually ended up matching MRI...so neat.

I also noticed we scoped first in the wrong way.  It is only
passed in for recursive array join in joinAny.  We do not use it
at all in the outer join function.  So I scoped it right before
the call which uses it.