No mode to_hash in any recv kwargs instrs.
This should never be needed since the only place this can
happen is in ruby2_keywords methods/hashes and that happens
on entry to any method/block now (for those cases).

The other thing added but currently used is whether we are
in a kwargs scope is being sent to mergeKeywordArguments.
This is unused but should be needed to prevent doing an
unconditional dup.  It is not hooked up yet because we have
to emit merge_kwargs differently in IRBuilder.  If a rest
kwarg is passed but used as a positional arg then we just use
the arg without duping.  Our IR will always construct an empty
hash first so this can never be true.  Unpeel the onion!