Tag out trivial error spec which errors already.
In order to support this and provide the proper error we need
to pass the instance of topself to classvarinstr instead of the
module of Object.  This is complicated because we pass in the module
for all accesses.  In most cases it is actually properly in a module
body so there is no instance to pass.