Fixes #6993.  Evals cause memory leak.
This makes two changes.  The first is to not record evals
as lexical children.  In IR we actually only use lexical
children for IR persistence (and AOT compilation).  We would
never persist a live eval instance.

This leads to the second change.  anyUsesEval was using lexicalChildren
to detect for evals when trying to perform AOT compilation (which we
do not allow).  This is better served by getClosures as we really
only care about "soft" scopes containing an eval (e.g. we don't
care if a method contains and eval if we are trying to AOT a module
scope (hard scope change) but we do care about a closure in a method
containing an eval (soft scope change)).