Only attempt reified super on Ruby < Java subclass
This logic always blindly proceeded to Ruby < Java subclass super
logic for all JavaProxy, but many JavaProxy are just normal Java
classes wrapped with our Ruby proxy logic. In this latter case, we
need to detect that it is not a Ruby < Java subclass (by the lack
of a JavaProxyClass) and dispatch through normal Java proxy
mechanisms.

Fixes #6968