Do not include .. into glob results.
This is partially a match to how MRI does this but our two impls
are fairly out of date with each other.  MRI is more clever and
uses . and .. in the file loop to determine whether we scan while
we essentially will check to see if we have . or .. later on again.