Update default and bundled gems for 3.1
Several gems have added JRuby support but we are stiil unable to
support many others. I have added issues and links for most gems
we do not currently support. I did not file issues for debug, rbs,
and typeprof because they are heavily vested in CRuby internals
and it is unclear how will will support them in the future.