Restore 9.2 collect signature and add more API
This addresses an issue found by jruby-openssl where we altered
the return value for a method in JRuby and broke backward compat
with JRuby 9.2 and earlier. Specifically we unintentionally caused
the RubyArray#collect(context, block) method to return IRubyObject
instead of RubyArray due to needing to return an enumerator for
some inputs.

As a result of this change, jruby-openssl compiled against the new
signature and became incompatible with JRuby 9.2.

This fix does not address the issue for JRuby 9.3.0 through 9.3.3,
which continue to have the bad signature. The restored signature
will remain through 9.3 and 9.4 maintenance at least, with two new
methods collectArray and collectEnum added to allow users to opt
into those specific paths. The Ruby-facing method is renamed to
rbCollect to indicate it is part of the Ruby interface for this
class.

See https://github.com/jruby/jruby-openssl/issues/245