Prefer the native thread status
The native status of a thread will be more accurate when the
thread is in a blocking or sleeping state, since we often cannot
detect these states before making calls into Java libraries. The
status we manage on our own is only set when calling through our
own blocking call wrappers, which are not present for blocking
calls made directly to Java.

The logic here uses the native status if it is anything other
than "new" or "run". Only if the native thread appears to be
runnable do we fall back on the Ruby thread status we maintain.

This appears to fix #7032 but it is unknown what other impact it
may have.