Tread LoadError as a completed load
LoadError should be treated as a completed load for locking
locking purposes, since it means an attempt was made to load the
file and that load completed with an error condition. This appears
to be different than the behavior for other errors that are raised
which should leave the require as being in-progress.