First pass of modernizing Marshal#load
The changes needed to support freeze: true uncovered a bunch of
other structural changes we needed to pass 11 past tagged out
issues (and 11 for restore since they are the same thing).

I have not run MRIs tests yet so I expect some more issues to appear.
This will pass all known ruby/spec issues except 3: one for T_OBJECT,
one for T_DATA, and one for a symbol which has multiple encodings but
the same bytes.  We do not have the concepts of T_DATA/T_OBJECT so
checking for them is not currently possible.  We also have a limitation
on symbols with the same bytes and multiple encodings.

There is still some cleanup work to be done.  We have some legacy methods
which if called may not work at all so I need to revisit than and hopefully
continue to support those signatures and still have things work.