Move native Timeout into Ruby
This largely works the same as the Java code, but written in Ruby.

It passes all but one test, test_skip_rescue, which requires the
timeout sub-thread to throw rather than raise the exception, so
that it can bypass rescue blocks and still be raised even if the
body of the timeout swallows the actual exception (throw is not
subject to rescues).