Always use ModuleSuper and rewrite define_method
The changes here attempt to get more super calls using a static
name, so we can eliminate dependency on a call frame to get that
name.

* super in a block in any method will start out as a ModuleSuper
  using the name from the method.
* define_method clones and rewrites the block to retarget any
  super calls to the newly defined name.
* all super instructions get rewritten to the appropriate type for
  the target class, instance/class/module

Remaining cases that use UnresolvedSuper may only be the ones
where super is invalid, such as a block at top-level or within a
class or module body.