First pass on adding ripper in same format as MRI
At this point we have the same lines as MRI which should be
enough to generate RipperParser directly.

The next part of this work is making RubyParser and RipperParser
close enough in how they extend and consume stuff that they can
both have the same implementations of common methods they both call.