More checking for null fd
This may resolve issues like #6669 where an in-use stream is
closed in another thread. We have some races here (which may be
unavoidable) that can lead to the fd field getting nulled out on
the way to a read or write. Without null checks, we may raise NPE
instead of a more appropriate EBADF when the fd is dereferenced.