Use non-range cover logic for eqq and include
This code has for years mistakenly dispatched to the full cover?
logic, which considers intersecting ranges as being covered. The
logic in CRuby, however, only uses a small part of the cover logic
that does not accept ranges. The naming of the methods (in CRuby
it's r_cover_p, and in JRuby that got mapped to the full cover_p
method instead of the equivalent rangeInclude method) probably
was confusing, or early in JRuby development CRuby's r_cover_p and
our cover_p were equivalent and that changed over time in CRuby.

This also eliminates a dynamic dispatch to `cover?` that does not
exist in CRuby.

Fixes #7137