Fixes #7102. ArgumentError re-splatting empty keyword arguments
Went with marking incoming key word rest arguments getting marked with
a flag indicating that it is a kwrest.  The flag is removed once we
enter check_arity helper.  A kwrest is transient and isolated from the
call site to the method which recieves it.  So there is no opportunity
for more than one thread to see this or other shenanigans like that.

This also is the general direction of Ruby 3 where we (and MRI) mark
kwargs hashes with a flag.  We seem to fail there as well so we might
actually need this new flag there (since kw and kwrest will have different
behavior when passed to a method like the one attached to #7102).