Redefine isBuiltin based on core boot
The original purpose of isBuiltin was to allow us to detect when
a dynamically-acquired method is actually the implementation we
shipped in a core JRuby class, to enable us to customize related
logic specific to how we know the JRuby implementation works. This
PR codifies that relationship, only setting methods to be
isBuiltin if they are constructed and bound during the core boot
process of JRuby itself.

In the future we will entertain more descriptive metadata for such
methods, but for now this narrows isBuiltin just enough to avoid
third-party extensions running into issues.

Fixes #7154