Fix random bignum path to dyncall #bytes
This logic in CRuby iteratively calls #bytes against the self obj
until it has the right amount of bits for the given bignum range.
Since we can construct a BigInteger from the bytes directly, we do
one call for the full octets needed and one additional call for
any additional bits.