closures were not being seen as ruby2_keywords.
This is moved to IRScope like it is for methods.  It fixes the immediate
problem.  Ultimately I am unsure if we should be storing this state on
IRScope at all.  In the case of closures we can have the same body in two
contexts.  One which would maybe be ruby2_keywords and one which would not.
Ultimately this is very rare but this is live state.  We should probably
move this field onto staticscope.