Native methods get extra argument from **{}.
When we build our call args we will merge kwargs and end up with {}
which we will push into a compound array.  Pure Ruby processing will
see we have an empty kwarg and remove it if not needed or use it if
there is a kwrest.  Native methods just see an extra argument.

There are two paths to go here:
1. Add this keyword logic to native methods via binding logic
2. Remove empty keywords since native methods have no concept of them
   and ruby methods will just make an empty one on demand if it is missing.

This took route number 2 and probably not in the best way.  When MERGE_KWARGS
occurs if it sees both arguments are {} then it will just return UNDEFINED.
This is a value which no one should ever see.  Luckily in the case of
keywords arguments this will end up hitting either args_push or args_cat
so we can filter out UNDEFINED in both places.

In looking at this I think we probably can be a bit smarter about what we emit.
We make lots of hash instances and likely with new instrs or a little more
examination we can make a simpler set of instrs do the same work.  This is
not trivial since when we work with variables or calls what we get back is
opaque to use during build-time.  Still it seems like we are making extra
hashes and arrays we may not need (since they are only for setting up args
lists).