What is old is new again.  Reinstate conditional kwrest dispatch.
For callsites which pass **kw we will check to see if it is empty
and then call a method with one less argument if so.  This is a little
undesirable to make an if/else for **kw sites but this is at the cost
of having to have all method prologues do this logic.