Array using **keysplat should not crash.
This is an outcome of **{} not really being a value.  In IR we will
make this undefined and cull it out of existence.  This in turn means
anything which can be composed or passed undefined must make an if/else
in IR to not pass it as a value.

Now that this is fixed I know hashes themselves (not used in calls) can
also do this.  So another commit shall be forthcoming.