Was unmarking ruby2_keyword last arguments too much.
Latest spec commit actually ended up providing a test for this.  This
is also echoed in active_module which has the same issue.  If you propate
a list of arguments through a series of methods which accept splats
(def foo(*args)).  It will never dup nor unmark the last argument
as a ruby2_keyword.