Add spec for JI interface dispatch with indy
This tests the fix for #7246. It must run twice because indy call
sites do not use the optimized handle for the first call through,
due to those handles being specialized to a specific signature and
the call site being a generic args[] call form. This is not ideal
in general, but this spec is sufficient to confirm the bug is
fixed until we can modify the indy call sites to always dispatch
using the optimized logic.