More forward logic...
We are halve-heartedly processing what I think will need to be a much
larger set of changes.  The issue is some native methods need to forward
their arguments to an eventual non-native method which will want to know
info like whether it is being passed keyword arguments.  This processing
is being done piece meal as we find cases but this probably needs to be
its own callpath for keyword arg dispatch.  We do have plans for this but
it was not going to happen before 9.4.0.0 is released.

In this commit for send/call/public_send we replicate that any calls
which pass **kw at the callsite to dup() those.