Eliminate JI-specific const replacement warning
This warning duplicates existing functionality intended to warn
when overwriting a constant, perhaps due to the more "magic"
nature of include_package and its open-ended behavior of assigning
constants to Java classes on demand. This feature is expected to
be used in places where you explicitly want the Java classes to be
preferred over classes further up in the hierarchy, so I believe
it makes sense to only warn if overwriting a constant that exists
on the target module directly.

Fixes #7185