Use the wide versions of these functions always
The two delete functions were modified to be wide in CRuby in
response to https://bugs.ruby-lang.org/issues/10820 seven years
ago. The query function was mentioned here but not updated to be
wide. I am unsure how CRuby is locating the appropriate `A` or `W`
function given the bare name, but since we have largely forked
this file and always want to use the wide version, this change
seems appropriate.

Fixes #7106