Fixes #7218. Empty rescue block returns exception object instead of nil.
The fix was an error in our parser.  When we get exception variables in
a rescue statement we will make effectively '_e = $!' and then append
whatever is actually in the body of the exception body.  In MRI's case
it ends up being 'nil'.  In our case it was 'null'.  Our impl thinks it
is smart and realizes nothing is happening and what should you append
if it is null anyways and just left the body as only '_e = $!'.  We
needed '_e = $!; nil'.

Trivial fix is to make sure any null return from compstmt in this grammar
production will always just be nil.  Impact from this change should only
effect the case when a rescue body is empty and in the precence of an
exception variable.  What we did was wrong.  I see no risk.