Only use NOSHARE_DELETE if accessible
The NOSHARE_DELETE ExtendedOption is not always available on all
systems, since it is part of a com.sun package that's not a public
API on recent JDKs. This commit guards its use, falling back on
the default NIO behavior of SHARE_DELETE when we can't access the
sole option that would turn it off.

Fixes #7268