Use null to make indy happy?
Not sure why but passing an explicit NullBlock operand will cause
indy to have an argument mismatch.  It treats NullBlock as an
explicit argument and it cannot bind IRubyObject, Block.

This could/should get fixed in indy impl but it was up until recently
null.  I just reverted this in IRBuilder.