Minor remove duplicate warnings.
We were printing this out twice per occurence:
warning: found `= literal' in conditional, should be ==

This was because parser used to call cond() on the value but with the
introduction of new_if (and also new_unless in MRI) it will perform this
from that new method.  Fix was to remove cond().