third argument cannot accept nil as a valid parameter.
This only happens unnaturally by intermediate methods which already
know the third argument is either really the options hash OR it was
not provided at all (varargs caller).  Replace using nil with null in
atOpts which upon seeing null will set the value to :microseconds.  If
it is nil it will blow up as expected.